+++ src/App.js
@@ -1,11 +1,11 @@
-import React, { useState } from "react";
+import React from "react";
 import "./styles.css";
-import { Routes, Route, Link } from 'react-router-dom';
-import { ProductListing } from "./Products";
-import { Cart } from "./Cart";
-import { Wishlist } from "./WishList";
-import NotFound from "./404";
-import Home from "./Home";
+import { Routes, Route} from 'react-router-dom';
+import { ProductListing } from "./Pages/Products";
+import { Cart } from "./Pages/Cart";
+import { Wishlist } from "./Pages/WishList";
+import NotFound from "./Pages/404";
+import Home from "./Pages/Home";
 import { NavigationBar } from "./Components/Navigation/navigationBar";
  
 export default function App() {
