+++ src/index.js
@@ -1,8 +1,8 @@
 import { StrictMode } from "react";
 import ReactDOM from "react-dom";
 import App from "./App";
-import { CartProvider } from "./cart-context";
-import { WishListProvider } from "./wishlist-context";
+import { CartProvider } from "./Context/cart-context";
+import { WishListProvider } from "./Context/wishlist-context";
 import { BrowserRouter as Router } from "react-router-dom";
 
 const rootElement = document.getElementById("root");
